+++ cpp/cudnn/include/layers/batch_norm_2d.cuh
@@ -11,19 +11,19 @@ class BatchNorm2d : public Layer {
 public:
     explicit BatchNorm2d(
         std::string name,
-        float epsilon = 1e-5,
-        float momentum = 0.1,
+        double epsilon = 1e-5,
+        double momentum = 0.1,
         bool affine = true,
         bool track_running_stats = true,
         cudnnBatchNormMode_t mode = CUDNN_BATCHNORM_SPATIAL_PERSISTENT);
     ~BatchNorm2d() override;
 
-    Tensor<float> *forward(Tensor<float> *input) override;
-    Tensor<float> *backward(Tensor<float> *grad_output) override;
+    Tensor<double> *forward(Tensor<double> *input) override;
+    Tensor<double> *backward(Tensor<double> *grad_output) override;
 
 private:
-    void fwd_initialize(Tensor<float> *input) override;
-    void bwd_initialize(Tensor<float> *grad_output) override;
+    void fwd_initialize(Tensor<double> *input) override;
+    void bwd_initialize(Tensor<double> *grad_output) override;
 
     size_t workspace_size_ = 0;
     size_t reserve_size_ = 0;
@@ -32,14 +32,14 @@ private:
     virtual void set_workspace();
 
     int num_features_;
-    float epsilon_;
-    float momentum_;
+    double epsilon_;
+    double momentum_;
     bool affine_;
     bool track_running_stats_;
-    Tensor<float> *running_mean_ = nullptr;
-    Tensor<float> *running_var_ = nullptr;
-    Tensor<float> *save_mean_ = nullptr;
-    Tensor<float> *save_var_ = nullptr;
+    Tensor<double> *running_mean_ = nullptr;
+    Tensor<double> *running_var_ = nullptr;
+    Tensor<double> *save_mean_ = nullptr;
+    Tensor<double> *save_var_ = nullptr;
 
     cudnnBatchNormMode_t mode_;
     cudnnTensorDescriptor_t derived_bn_scale_bias_mean_var_desc_;
