+++ cpp/cudnn/include/layers/conv_2d.cuh
@@ -23,12 +23,12 @@ public:
         bool bias = true);
     ~Conv2d() override;
 
-    Tensor<float> *forward(Tensor<float> *input) override;
-    Tensor<float> *backward(Tensor<float> *grad_output) override;
+    Tensor<double> *forward(Tensor<double> *input) override;
+    Tensor<double> *backward(Tensor<double> *grad_output) override;
 
 private:
-    void fwd_initialize(Tensor<float> *input) override;
-    void bwd_initialize(Tensor<float> *grad_output) override;
+    void fwd_initialize(Tensor<double> *input) override;
+    void bwd_initialize(Tensor<double> *grad_output) override;
 
     int out_channels_;
     int kernel_size_;
